Breaking cycle (RecordHeader to init.py) + wrapping SSS
