+++ Mopy/bash/brec.py
@@ -26,7 +26,6 @@
 import zlib
 import StringIO
 import os
-import re # flagged as unused but game detection fails if deleted
 import struct
 import copy
 import cPickle
