+++ Mopy/bash/game/skyrim/__init__.py
@@ -24,10 +24,14 @@
 
 """This modules defines static data for use by bush, when TES V:
    Skyrim is set at the active game."""
-
+import re
+import struct
 from constants import bethDataFiles, allBethFiles
 from ... import brec
-from ...brec import *
+from records import MreAact, MreActi, MreAddn, MreAmmo, MreAnio, MreAppa, \
+    MreArma, MreArmo, MreArto, MreAspc, MreAstp, MreCobj, MreGmst, MreLvli, \
+    MreLvln, MreLvsp, MreMisc, MreHeader
+from ...brec import MreGlob, BaseRecordHeader, ModError
 
 #--Name of the game to use in UI.
 displayName = u'Skyrim'
@@ -1379,28 +1383,89 @@ class esp:
         ]
 
     #--Top types in Skyrim order.
-    topTypes = ['GMST', 'KYWD', 'LCRT', 'AACT', 'TXST', 'GLOB', 'CLAS', 'FACT', 'HDPT',
-                'HAIR', 'EYES', 'RACE', 'SOUN', 'ASPC', 'MGEF', 'SCPT', 'LTEX', 'ENCH',
-                'SPEL', 'SCRL', 'ACTI', 'TACT', 'ARMO', 'BOOK', 'CONT', 'DOOR', 'INGR',
-                'LIGH', 'MISC', 'APPA', 'STAT', 'SCOL', 'MSTT', 'PWAT', 'GRAS', 'TREE',
-                'CLDC', 'FLOR', 'FURN', 'WEAP', 'AMMO', 'NPC_', 'LVLN', 'KEYM', 'ALCH',
-                'IDLM', 'COBJ', 'PROJ', 'HAZD', 'SLGM', 'LVLI', 'WTHR', 'CLMT', 'SPGD',
-                'RFCT', 'REGN', 'NAVI', 'CELL', 'WRLD', 'DIAL', 'QUST', 'IDLE', 'PACK',
-                'CSTY', 'LSCR', 'LVSP', 'ANIO', 'WATR', 'EFSH', 'EXPL', 'DEBR', 'IMGS',
-                'IMAD', 'FLST', 'PERK', 'BPTD', 'ADDN', 'AVIF', 'CAMS', 'CPTH', 'VTYP',
-                'MATT', 'IPCT', 'IPDS', 'ARMA', 'ECZN', 'LCTN', 'MESG', 'RGDL', 'DOBJ',
-                'LGTM', 'MUSC', 'FSTP', 'FSTS', 'SMBN', 'SMQN', 'SMEN', 'DLBR', 'MUST',
-                'DLVW', 'WOOP', 'SHOU', 'EQUP', 'RELA', 'SCEN', 'ASTP', 'OTFT', 'ARTO',
-                'MATO', 'MOVT', 'SNDR', 'DUAL', 'SNCT', 'SOPM', 'COLL', 'CLFM', 'REVB',]
+    topTypes = ['GMST', 'KYWD', 'LCRT', 'AACT', 'TXST', 'GLOB', 'CLAS', 'FACT',
+                'HDPT', 'HAIR', 'EYES', 'RACE', 'SOUN', 'ASPC', 'MGEF', 'SCPT',
+                'LTEX', 'ENCH', 'SPEL', 'SCRL', 'ACTI', 'TACT', 'ARMO', 'BOOK',
+                'CONT', 'DOOR', 'INGR', 'LIGH', 'MISC', 'APPA', 'STAT', 'SCOL',
+                'MSTT', 'PWAT', 'GRAS', 'TREE', 'CLDC', 'FLOR', 'FURN', 'WEAP',
+                'AMMO', 'NPC_', 'LVLN', 'KEYM', 'ALCH', 'IDLM', 'COBJ', 'PROJ',
+                'HAZD', 'SLGM', 'LVLI', 'WTHR', 'CLMT', 'SPGD', 'RFCT', 'REGN',
+                'NAVI', 'CELL', 'WRLD', 'DIAL', 'QUST', 'IDLE', 'PACK', 'CSTY',
+                'LSCR', 'LVSP', 'ANIO', 'WATR', 'EFSH', 'EXPL', 'DEBR', 'IMGS',
+                'IMAD', 'FLST', 'PERK', 'BPTD', 'ADDN', 'AVIF', 'CAMS', 'CPTH',
+                'VTYP', 'MATT', 'IPCT', 'IPDS', 'ARMA', 'ECZN', 'LCTN', 'MESG',
+                'RGDL', 'DOBJ', 'LGTM', 'MUSC', 'FSTP', 'FSTS', 'SMBN', 'SMQN',
+                'SMEN', 'DLBR', 'MUST', 'DLVW', 'WOOP', 'SHOU', 'EQUP', 'RELA',
+                'SCEN', 'ASTP', 'OTFT', 'ARTO', 'MATO', 'MOVT', 'SNDR', 'DUAL',
+                'SNCT', 'SOPM', 'COLL', 'CLFM', 'REVB', ]
 
     #--Dict mapping 'ignored' top types to un-ignored top types.
-    topIgTypes = dict([(struct.pack('I',(struct.unpack('I',type)[0]) | 0x1000),type) for type in topTypes])
+    topIgTypes = dict(
+        [(struct.pack('I', (struct.unpack('I', type)[0]) | 0x1000), type) for
+         type in topTypes])
 
     #-> this needs updating for Skyrim
-    recordTypes = set(topTypes + 'GRUP,TES4,REFR,ACHR,ACRE,LAND,INFO,NAVM,PHZD,PGRE'.split(','))
+    recordTypes = set(
+        topTypes + 'GRUP,TES4,REFR,ACHR,ACRE,LAND,INFO,NAVM,PHZD,PGRE'.split(
+            ','))
 
-#------------------------------------------------------------------------------
-from records import * # MUST BE HERE otherwise all hell breaks loose
+#--Mod I/O
+class RecordHeader(BaseRecordHeader):
+    size = 24
+
+    def __init__(self,recType='TES4',size=0,arg1=0,arg2=0,arg3=0,extra=0):
+        self.recType = recType
+        self.size = size
+        if recType == 'GRUP':
+            self.label = arg1
+            self.groupType = arg2
+            self.stamp = arg3
+        else:
+            self.flags1 = arg1
+            self.fid = arg2
+            self.flags2 = arg3
+        self.extra = extra
+
+    @staticmethod
+    def unpack(ins):
+        """Returns a RecordHeader object by reading the input stream."""
+        type,size,uint0,uint1,uint2,uint3 = ins.unpack('=4s5I',24,'REC_HEADER')
+        #--Bad type?
+        if type not in esp.recordTypes:
+            raise ModError(ins.inName,u'Bad header type: '+repr(type))
+        #--Record
+        if type != 'GRUP':
+            pass
+        #--Top Group
+        elif uint1 == 0: #groupType == 0 (Top Type)
+            str0 = struct.pack('I',uint0)
+            if str0 in esp.topTypes:
+                uint0 = str0
+            elif str0 in esp.topIgTypes:
+                uint0 = esp.topIgTypes[str0]
+            else:
+                raise ModError(ins.inName,u'Bad Top GRUP type: '+repr(str0))
+        #--Other groups
+        return RecordHeader(type,size,uint0,uint1,uint2,uint3)
+
+    def pack(self):
+        """Return the record header packed into a bitstream to be written to file."""
+        if self.recType == 'GRUP':
+            if isinstance(self.label,str):
+                return struct.pack('=4sI4sIII',self.recType,self.size,
+                                   self.label,self.groupType,self.stamp,
+                                   self.extra)
+            elif isinstance(self.label,tuple):
+                return struct.pack('=4sIhhIII',self.recType,self.size,
+                                   self.label[0],self.label[1],self.groupType,
+                                   self.stamp,self.extra)
+            else:
+                return struct.pack('=4s5I',self.recType,self.size,self.label,
+                                   self.groupType,self.stamp,self.extra)
+        else:
+            return struct.pack('=4s5I',self.recType,self.size,self.flags1,
+                               self.fid,self.flags2,self.extra)
+
 #------------------------------------------------------------------------------
 # Unused records, they have empty GRUP in skyrim.esm---------------------------
 # CLDC ------------------------------------------------------------------------
